pyproject.toml
[project]
name = "loremaster"
version = "0.1.0"
description = "A small agentic AI workflow for depicting fictional characters."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.4.1",
    "crewai-tools>=1.4.1",
    "google-genai>=1.50.1",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "typer>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["files_router"]

[tool.uv]
package = false
default-groups = "all"

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "mypy>=1.18.2",
    "ruff>=0.14.5",
]

[tool.mypy]
files = ["."]
ignore_missing_imports = true
check_untyped_defs = true

[tool.ruff]
line-length = 88
output-format = "grouped"

[tool.ruff.lint]
# see https://docs.astral.sh/ruff/settings/#lint
select = [  # see https://docs.astral.sh/ruff/rules/
    "E4", "E7", "E9",  # basic pycodestyle rules
    "F",        # Pyflakes rules
    "W",        # pycodestyle warnings
    "C90",      # McCabe complexity
    "N",        # PEP8 naming conventions
    "I",        # import statements linting and formatting
    "PLR",      # Pylint refactoring recommendations
    "PTH",      # recommends to use pathlib instead of os.path
    "UP",       # pyupgrade suggestions
    "BLE",      # locates blind exception capture
    "B",        # locates possible bugs and design problems
    "A",        # locates shadowing of builtin names
    "C4",       # locates possible issues with comprehensions
    "DTZ",      # locates incorrect use of timezone information
    "PIE",      # miscellaneous linting rules
    "TRY",      # locates potential exception handling antipatterns
    "PERF",     # locates potential performance issues
    "RUF",      # various Ruff-specific linting rules
    "S",        # various insecure code, with bandit
    "ERA",      # eradicate: commented-out code
]
ignore = [
    "ERA001",   # allow commented our code, for now
]
external = [
    "F841",     # vulture uses this code for ignoring unused names
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",     # allow unused imports
    "F403",     # allow `import *`
]
"tests/*" = [
    "PLR0913",  # unlimited function arguments
    "S101",     # allow asserts
    "S105",     # allow harcoded passwords
    "N802",     # allow variable case in test names
    "PLR2004",  # allow magic values in comparisons
    "S311"      # allow random generator
]

[tool.ruff.lint.isort]
# see https://docs.astral.sh/ruff/settings/#lintisort
#known-first-party = ["files_router"]
force-sort-within-sections = true

[tool.ruff.format]
# see https://docs.astral.sh/ruff/settings/#format

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

[tool.deptry.per_rule_ignores]
DEP002 = [
    "google-genai",
    "crewai-tools",
    "pillow",
]
